+++ src/main/java/com/oroarmor/multiitemlib/mixin/EfficiencyEnchantmentMixin.java
@@ -22,21 +22,21 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin.render;
+package com.oroarmor.multiitemlib.mixin;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.client.render.entity.FishingBobberEntityRenderer;
+import net.minecraft.enchantment.EfficiencyEnchantment;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-@Mixin(FishingBobberEntityRenderer.class)
-public class FishingBobberEntityRendererMixin {
-    @Redirect(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
-    private boolean render(ItemStack stack, Item isOfItem) {
-        return UniqueItemRegistry.FISHING_ROD.isItemInRegistry(stack.getItem());
+@Mixin(EfficiencyEnchantment.class)
+public class EfficiencyEnchantmentMixin {
+    @Redirect(method = "isAcceptableItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean isAcceptableItem(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.SHEARS.isItemInRegistry(stack.getItem());
     }
 }