+++ src/main/java/com/oroarmor/multiitemlib/mixin/EntityInteractShearsMixin.java
@@ -22,21 +22,22 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin;
+package com.oroarmor.multiitemlib.mixin;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.enchantment.UnbreakingEnchantment;
+import net.minecraft.entity.passive.SheepEntity;
+import net.minecraft.entity.passive.SnowGolemEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-@Mixin(UnbreakingEnchantment.class)
-public class UnbreakingEnchantmentMixin {
-    @Redirect(method = "shouldPreventDamage", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;getItem()Lnet/minecraft/item/Item;"))
-    private static Item shouldPreventDamage(ItemStack stack) {
-        return UniqueItemRegistry.ELYTRA.getDefaultItem(stack.getItem());
+@Mixin({SheepEntity.class, SnowGolemEntity.class})
+public class EntityInteractShearsMixin {
+    @Redirect(method = "interactMob", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean interactMob(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.SHEARS.isItemInRegistry(stack.getItem());
     }
 }