+++ src/main/java/com/oroarmor/multiitemlib/mixin/FishBobberEntityMixin.java
@@ -22,21 +22,21 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin.render;
+package com.oroarmor.multiitemlib.mixin;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.client.render.entity.feature.CapeFeatureRenderer;
+import net.minecraft.entity.projectile.FishingBobberEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-@Mixin(CapeFeatureRenderer.class)
-public class CapeFeatureRendererMixin {
-    @Redirect(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
-    private boolean render(ItemStack stack, Item isOfItem) {
-        return UniqueItemRegistry.ELYTRA.isItemInRegistry(stack.getItem());
+@Mixin(FishingBobberEntity.class)
+public class FishBobberEntityMixin {
+    @Redirect(method = "removeIfInvalid", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean removeIfInvalid(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.FISHING_ROD.isItemInRegistry(stack.getItem());
     }
 }