+++ src/main/java/com/oroarmor/multiitemlib/mixin/PlayerEntityMixin.java
@@ -22,21 +22,26 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin;
+package com.oroarmor.multiitemlib.mixin;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-@Mixin(LivingEntity.class)
-public class LivingEntityMixin {
-    @Redirect(method = "tickFallFlying()V", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
-    private boolean tickFallFlying(ItemStack stack, Item isOfItem) {
+@Mixin(PlayerEntity.class)
+public class PlayerEntityMixin {
+    @Redirect(method = "checkFallFlying()Z", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean checkFallFlying(ItemStack stack, Item isOfItem) {
         return UniqueItemRegistry.ELYTRA.isItemInRegistry(stack.getItem());
+    }
+
+    @Redirect(method = "damageShield(F)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean shields(ItemStack stack, Item isOfITem) {
+        return UniqueItemRegistry.SHIELD.isItemInRegistry(stack.getItem());
     }
 }