+++ src/main/java/com/oroarmor/multiitemlib/mixin/PumpkinBlockMixin.java
@@ -22,21 +22,21 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin;
+package com.oroarmor.multiitemlib.mixin;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.entity.passive.MooshroomEntity;
+import net.minecraft.block.PumpkinBlock;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-@Mixin({MooshroomEntity.class})
-public class MooshroomEntityInteractShearsMixin {
-    @Redirect(method = "interactMob", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z", ordinal = 1))
-    private boolean interactMob(ItemStack stack, Item isOfItem) {
-        return UniqueItemRegistry.ELYTRA.isItemInRegistry(stack.getItem());
+@Mixin(PumpkinBlock.class)
+public class PumpkinBlockMixin {
+    @Redirect(method = "onUse", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean onUse(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.SHEARS.isItemInRegistry(stack.getItem());
     }
 }