+++ src/main/java/com/oroarmor/multiitemlib/mixin/render/CapeFeatureRendererMixin.java
@@ -22,21 +22,21 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin;
+package com.oroarmor.multiitemlib.mixin.render;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.block.TripwireBlock;
+import net.minecraft.client.render.entity.feature.CapeFeatureRenderer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
-@Mixin(TripwireBlock.class)
-public class TripwireBlockMixin {
-    @Redirect(method = "onBreak", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
-    private boolean onBreak(ItemStack stack, Item isOfItem) {
-        return UniqueItemRegistry.SHEARS.isItemInRegistry(stack.getItem());
+@Mixin(CapeFeatureRenderer.class)
+public class CapeFeatureRendererMixin {
+    @Redirect(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean render(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.ELYTRA.isItemInRegistry(stack.getItem());
     }
 }