+++ src/main/java/com/oroarmor/multiitemlib/mixin/render/FishingBobberEntityRendererMixin.java
@@ -22,30 +22,21 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin.render;
+package com.oroarmor.multiitemlib.mixin.render;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.client.render.item.HeldItemRenderer;
+import net.minecraft.client.render.entity.FishingBobberEntityRenderer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-
-@Mixin(HeldItemRenderer.class)
-public class HeldItemRendererMixin {
-    @Redirect(method = "renderFirstPersonItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z", ordinal = 2))
-    private boolean renderFirstPersonItem(ItemStack stack, Item isOfItem) {
-        return UniqueItemRegistry.CROSSBOW.isItemInRegistry(stack.getItem());
-    }
-
-    @Redirect(method = "getHandRenderType", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
-    private static boolean getHandRenderType(ItemStack stack, Item isOfItem) {
-        return UniqueItemRegistry.BOW.isItemInRegistry(stack.getItem()) || UniqueItemRegistry.CROSSBOW.isItemInRegistry(stack.getItem());
+@Mixin(FishingBobberEntityRenderer.class)
+public class FishingBobberEntityRendererMixin {
+    @Redirect(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean render(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.FISHING_ROD.isItemInRegistry(stack.getItem());
     }
 }