+++ src/main/java/com/oroarmor/multiitemlib/mixin/render/HeldItemRendererMixin.java
@@ -22,31 +22,26 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin.render;
+package com.oroarmor.multiitemlib.mixin.render;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.client.network.AbstractClientPlayerEntity;
-import net.minecraft.client.render.entity.PlayerEntityRenderer;
-import net.minecraft.client.render.entity.model.BipedEntityModel;
-import net.minecraft.item.CrossbowItem;
+import net.minecraft.client.render.item.HeldItemRenderer;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.Hand;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
+@Mixin(HeldItemRenderer.class)
+public class HeldItemRendererMixin {
+    @Redirect(method = "renderFirstPersonItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z", ordinal = 1))
+    private boolean renderFirstPersonItem(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.CROSSBOW.isItemInRegistry(stack.getItem());
+    }
 
-@Mixin(PlayerEntityRenderer.class)
-public class PlayerEntityRendererMixin {
-    @Inject(method = "getArmPose", at = @At("HEAD"), cancellable = true)
-    private static void getArmPose(AbstractClientPlayerEntity abstractClientPlayerEntity, Hand hand, CallbackInfoReturnable<BipedEntityModel.ArmPose> cir) {
-        ItemStack itemStack = abstractClientPlayerEntity.getStackInHand(hand);
-        if (!abstractClientPlayerEntity.handSwinging && UniqueItemRegistry.CROSSBOW.isItemInRegistry(itemStack.getItem()) && CrossbowItem.isCharged(itemStack)) {
-            cir.setReturnValue(BipedEntityModel.ArmPose.CROSSBOW_HOLD);
-        }
+    @Redirect(method = "getHandRenderType", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private static boolean getHandRenderType(ItemStack stack, Item isOfItem) {
+        return UniqueItemRegistry.BOW.isItemInRegistry(stack.getItem()) || UniqueItemRegistry.CROSSBOW.isItemInRegistry(stack.getItem());
     }
 }