+++ src/main/java/com/oroarmor/multiitemlib/mixin/render/PlayerEntityRendererMixin.java
@@ -22,21 +22,28 @@
  * SOFTWARE.
  */
 
-package com.oroarmor.multi_item_lib.mixin.render;
+package com.oroarmor.multiitemlib.mixin.render;
 
-import com.oroarmor.multi_item_lib.UniqueItemRegistry;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import net.minecraft.client.render.item.ItemRenderer;
-import net.minecraft.item.Item;
+import net.minecraft.client.network.AbstractClientPlayerEntity;
+import net.minecraft.client.render.entity.PlayerEntityRenderer;
+import net.minecraft.client.render.entity.model.BipedEntityModel;
+import net.minecraft.item.CrossbowItem;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.Hand;
 
-@Mixin(ItemRenderer.class)
-public abstract class ItemRendererMixin {
-    @Redirect(method = "getHeldItemModel", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z", ordinal = 1))
-    private static boolean getHeldItemModel(ItemStack stack, Item isOfItem) {
-        return UniqueItemRegistry.TRIDENT.isItemInRegistry(stack.getItem());
+@Mixin(PlayerEntityRenderer.class)
+public class PlayerEntityRendererMixin {
+    @Inject(method = "getArmPose", at = @At("HEAD"), cancellable = true)
+    private static void getArmPose(AbstractClientPlayerEntity abstractClientPlayerEntity, Hand hand, CallbackInfoReturnable<BipedEntityModel.ArmPose> cir) {
+        ItemStack itemStack = abstractClientPlayerEntity.getStackInHand(hand);
+        if (!abstractClientPlayerEntity.handSwinging && UniqueItemRegistry.CROSSBOW.isItemInRegistry(itemStack.getItem()) && CrossbowItem.isCharged(itemStack)) {
+            cir.setReturnValue(BipedEntityModel.ArmPose.CROSSBOW_HOLD);
+        }
     }
 }