+++ CHANGELOG.md
@@ -1,12 +1,26 @@
+Version 1.4.0
+
+- Fix mixin crash issue in production
+- Move main package from `com.oroarmor.multi_item_lib` to `com.oroarmor.multiitemlib`
+- Moved `UniqueItemRegistry` into `api` sub-package.
+- Old `UniqueItemRegistry` is now deprecated
+
+----
 Version 1.3.2
+
 - Update to Minecraft 1.17
+
 ----
 Version 1.3.1
+
 - Javadocs
+
 ----
 Version 1.3.0
+
 - Add shears
 - Publish to Maven Central instead of Bintray
+
 ----
 Version 1.2.0
 - Add mixin for fishing bobber